+++ test/plugin_functional/test_suites/core_plugins/rendering.ts
@@ -128,8 +128,6 @@ export default function ({ getService }: PluginFunctionalProviderContext) {
         'data_visualizer.resultLinks.fileBeat.enabled (boolean)',
         'dev_tools.deeplinks.navLinkStatus (string?)',
         'discover.experimental.enabledProfiles (array?)',
-        'enterpriseSearch.host (string?)',
-        'enterpriseSearch.ui.enabled (boolean?)',
         'home.disableWelcomeScreen (boolean?)',
         'management.deeplinks.navLinkStatus (string?)',
         'map.emsFileApiUrl (string?)',
