+++ x-pack/solutions/search/plugins/enterprise_search/public/applications/test_helpers/test_utils.test_helper.tsx
@@ -45,7 +45,6 @@ export const mockKibanaProps: KibanaLogicProps = {
     isCloudEnabled: false,
   },
   config: {
-    host: 'http://localhost:3002',
     ui: {
       enabled: true,
     },
